// src/proto/comment/rpc_get_post_comments.ts
export interface GetPostCommentsRequest {
    /**
     * @generated from protobuf field: string post_id = 1;
     */
    postId: string;
    /**
     * @generated from protobuf field: int32 per_page = 2;
     */
    perPage: number;
    /**
     * @generated from protobuf field: int32 page = 3;
     */
    page: number;
}
// ...
class GetPostCommentsRequest$Type extends MessageType<GetPostCommentsRequest> {
    constructor() {
        super("pb.GetPostCommentsRequest", [
            { no: 1, name: "post_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "per_page", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "page", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
    create(value?: PartialMessage<GetPostCommentsRequest>): GetPostCommentsRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.postId = "";
        message.perPage = 0;
        message.page = 0;
        if (value !== undefined)
            reflectionMergePartial<GetPostCommentsRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetPostCommentsRequest): GetPostCommentsRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string post_id */ 1:
                    message.postId = reader.string();
                    break;
                case /* int32 per_page */ 2:
                    message.perPage = reader.int32();
                    break;
                case /* int32 page */ 3:
                    message.page = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GetPostCommentsRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string post_id = 1; */
        if (message.postId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.postId);
        /* int32 per_page = 2; */
        if (message.perPage !== 0)
            writer.tag(2, WireType.Varint).int32(message.perPage);
        /* int32 page = 3; */
        if (message.page !== 0)
            writer.tag(3, WireType.Varint).int32(message.page);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message pb.GetPostCommentsRequest
 */
export const GetPostCommentsRequest = new GetPostCommentsRequest$Type();