// src/proto/comment/rpc_create_comment.ts
export interface CreateCommentRequest {
    /**
     * @generated from protobuf field: string post_id = 1;
     */
    postId: string;
    /**
     * @generated from protobuf field: string text = 2;
     */
    text: string;
}
/**
 * @generated from protobuf message pb.CreateCommentResponse
 */
export interface CreateCommentResponse {
    /**
     * @generated from protobuf field: pb.Comment comment = 1;
     */
    comment?: Comment;
}
// @generated message type with reflection information, may provide speed optimized methods
class CreateCommentRequest$Type extends MessageType<CreateCommentRequest> {
    constructor() {
        super("pb.CreateCommentRequest", [
            { no: 1, name: "post_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "text", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<CreateCommentRequest>): CreateCommentRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.postId = "";
        message.text = "";
        if (value !== undefined)
            reflectionMergePartial<CreateCommentRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: CreateCommentRequest): CreateCommentRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string post_id */ 1:
                    message.postId = reader.string();
                    break;
                case /* string text */ 2:
                    message.text = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: CreateCommentRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string post_id = 1; */
        if (message.postId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.postId);
        /* string text = 2; */
        if (message.text !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.text);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message pb.CreateCommentRequest
 */
export const CreateCommentRequest = new CreateCommentRequest$Type();